id type changed

+++ Task4/AppCore/ProjectStructure.Domain/Ticket.cs
@@ -8,6 +8,6 @@ namespace ProjectStructure.Domain
     {
         //public int Id { get; set; }
         public double Price { get; set; }
-        public int FlightId { get; set; }
+        public long FlightId { get; set; }
     }
 }
